Store spotify auth expiration as Long instead of Int to prevent overflow in year 2038

+++ app/src/main/java/io/libzy/spotify/api/SpotifyApiDelegator.kt
@@ -60,7 +60,7 @@ class SpotifyApiDelegator @Inject constructor(
     private suspend fun createApiDelegateIfTokenAvailable(): SpotifyClientApi? {
         with(context.getSharedPrefs()) {
             val savedAccessToken = getString(SharedPrefKeys.SPOTIFY_AUTH_TOKEN, null)
-            val savedTokenExpiration = getInt(SharedPrefKeys.SPOTIFY_AUTH_EXPIRATION, 0)
+            val savedTokenExpiration = getLong(SharedPrefKeys.SPOTIFY_AUTH_EXPIRATION, 0)
             if (savedAccessToken != null && currentTimeSeconds() < savedTokenExpiration) {
                 _apiDelegate = createApiDelegate(savedAccessToken)
             }
